fix: rename to ts file
