+++ src/templates/next/components/Infobar/Infobar.tsx
@@ -1,6 +1,6 @@
 import { InfobarProps } from "~/common"
 import Button from "../Button"
-import { ComponentContent } from "../shared/CustomCssClass"
+import { ComponentContent } from "../shared/customCssClass"
 
 const Infobar = ({
   sectionIdx,
